Move the computation of normals to reduce redundant computation.
Before, normals were computed as triangles were pushed to the indices
array. However, this caused the same normals to be needlessly recomputed
(up to six times, half of which we cannot avoid without a major
overhaul).
By moving the normal calculation to occur at the same time as the
vertices are added to the buffer, most of the redundant normal
computations have been removed.